requirements.txt
numpy
matplotlib
joblib
cvxpy
pandas
pyyaml
scipy
seaborn